+++ spec/models/permitted_params_spec.rb
@@ -711,4 +711,38 @@ describe PermittedParams do
       it_behaves_like 'allows params'
     end
   end
+
+  describe 'member' do
+    let (:attribute) { :member }
+
+    describe 'role_ids' do
+      let(:hash) { { "role_ids" => [] } }
+
+      it_behaves_like 'allows params'
+    end
+
+    describe 'user_id' do
+      let(:hash) { { "user_id" => "blubs" } }
+
+      it_behaves_like 'forbids params'
+    end
+
+    describe 'project_id' do
+      let(:hash) { { "user_id" => "blubs" } }
+
+      it_behaves_like 'forbids params'
+    end
+
+    describe 'created_on' do
+      let(:hash) { { "created_on" => "blubs" } }
+
+      it_behaves_like 'forbids params'
+    end
+
+    describe 'mail_notification' do
+      let(:hash) { { "mail_notification" => "blubs" } }
+
+      it_behaves_like 'forbids params'
+    end
+  end
 end